Break out previousEvents to its own store so it can be imported
