+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@temporalio/ui",
-  "version": "2.1.71",
+  "version": "2.1.72",
   "type": "module",
   "description": "Temporal.io UI",
   "keywords": [
