+++ src/lib/stores/previous-events.ts
@@ -1,0 +1,18 @@
+import { writable, Writable } from 'svelte/store';
+
+import type { FetchEventsParameters } from '$lib/services/events-service';
+
+const emptyPrevious: FetchEventsParameters = {
+  namespace: null,
+  workflowId: null,
+  runId: null,
+  rawPayloads: null,
+  sort: null,
+};
+
+export const previous: Writable<FetchEventsParameters> =
+  writable(emptyPrevious);
+
+export const clearPreviousEventParameters = (): void => {
+  previous.set(emptyPrevious);
+};
