+++ src/routes/namespaces/[namespace]/workflows/[workflow]/[run]/__layout@root.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import WorkflowRunLayout from '$lib/layouts/workflow-run-layout.svelte';
   import { onDestroy } from 'svelte';
-  import { clearPreviousEventParameters } from '$lib/stores/events';
+  import { clearPreviousEventParameters } from '$lib/stores/previous-events';
 
   onDestroy(() => {
     clearPreviousEventParameters();
